// apps/client/src/app/charteditor/utils/events.ts
export enum EVENTS {
  COPY_TO_CLIPBAORD = 'copy-to-clipboard',
  EXPORT_TO_PNG = 'export-to-image',
  EXPORT_TO_PDF = 'export-to-pdf',
  EXPORT_TO_SVG = 'export-to-svg',
  EXPORT_TO_JPG = 'export-to-jpg',
  EMBED_STATIC_IMAGE = 'embed-static-image',
  UPLOAD_EMBED_STATIC_IMAGE = 'upload-embed-static-image',
}